[BE] - ✨Feature: 단기 예보 조회 배치 서버 구현

+++ backend/application-modules/batch-server/src/main/java/com/softeer/batch/forecast/mountain/dto/MountainDailyForecast.java
@@ -8,14 +8,14 @@ import java.util.List;
 
 public record MountainDailyForecast(
         long mountainId,
-        long gridId,
+        int gridId,
         SunTime sunTime,
         List<Forecast> hourlyForecasts
 ) {
 
     public MountainDailyForecast(
             long mountainId,
-            long gridId,
+            int gridId,
             LocalTime sunrise,
             LocalTime sunset,
             List<Forecast> hourlyForecasts
