+++ backend/application-modules/batch-server/src/main/java/com/softeer/batch/forecast/mountain/dto/MountainIdentifier.java
@@ -3,6 +3,6 @@ package com.softeer.batch.forecast.mountain.dto;
 public record MountainIdentifier(
         long id,
         int code,
-        long gridId
+        int gridId
 ) {
 }
